// src/kavak/interfaces/interfaces.ts
export interface MyDataResponse {
    message: string;
    data: MyData   
}

export interface MyData {
    idkavakpoc: string;
    tvplaca?: string | null;
    tvmes?: string | null;
    tvestado?: string | null;
    tcplaca?: string | null;
    tcvin?: string | null;
    tcmodelo?: string | null;
    tcestado?: string | null;
    cftipo?: string | null;
    cfplaca?: string | null;
    cdmodelo?: string | null;
    cfanio?: string | null;
    cfvalidez?: string | null;
    cfcodigo?: string | null;
    paclave?: string | null;
    peclave?: string | null;
    c5docs?: string | null;
    reclave?: string | null;
    vinparabclave?: string | null;
    vincarclave?: string | null;
    vinmotorclave?: string | null;
    vinescannerclave?: string | null;
    coicniden4docs?: string | null;
    kmtacometroclave?: string | null;
    kmescannerclave?: string | null;
    coiciden2docs?: string | null;
    validacionrepuve?: string | null;
    repuvevin?: string | null;
    repuveanio?: string | null;
    repuvemarca?: string | null;
    repuvemodelo?: string | null;
    repuveprocedencia?: string | null;
    documentosvin?: string | null;
    documentosanio?: string | null;
    documentosmarca?: string | null;
    documenrosmodelo?: string | null;
    documentosprocedencia?: string | null;
    FGJ?: string | null;
    OCRA?: string | null;
    USA7CAN?: string | null;
    AVISOSJUDICALES?: string | null;
    VALIDACIONRAPI?: string | null;
    RAPIVIN?: string | null;
    VALIDACIONAMIREPORTEROBO?: string | null;
    VALIDACIONamiSEGUROS?: string | null;
    AMIVIGENCIA?: string | null;
    AMIPOLIZA?: string | null;
    AMIMARCA?: string | null;
    ADEUDOS?: string | null;
    ADEUDOEDOMEX?: string | null;
    ADEUDOCDMX?: string | null;
    tcfechavigencia?: string | null;
    tcedovigente?: string | null;
    tcmotor?: string | null;
  }
  
  